+++ include/bimtree/PseudoPointGenerator.hpp
@@ -21,55 +21,55 @@
 namespace icp {
 
     /// Calculate triangle area using cross product
-    inline float triangleArea(
+    inline double triangleArea(
         const pctree::XYZPoint& v0,
         const pctree::XYZPoint& v1,
         const pctree::XYZPoint& v2)
     {
         /// Edge vectors
-        float e1x = v1[0] - v0[0];
-        float e1y = v1[1] - v0[1];
-        float e1z = v1[2] - v0[2];
+        double e1x = v1[0] - v0[0];
+        double e1y = v1[1] - v0[1];
+        double e1z = v1[2] - v0[2];
 
-        float e2x = v2[0] - v0[0];
-        float e2y = v2[1] - v0[1];
-        float e2z = v2[2] - v0[2];
+        double e2x = v2[0] - v0[0];
+        double e2y = v2[1] - v0[1];
+        double e2z = v2[2] - v0[2];
 
         /// Cross product
-        float cx = e1y * e2z - e1z * e2y;
-        float cy = e1z * e2x - e1x * e2z;
-        float cz = e1x * e2y - e1y * e2x;
+        double cx = e1y * e2z - e1z * e2y;
+        double cy = e1z * e2x - e1x * e2z;
+        double cz = e1x * e2y - e1y * e2x;
 
         /// Area = 0.5 * |cross product|
-        return 0.5f * std::sqrt(cx * cx + cy * cy + cz * cz);
+        return 0.5 * std::sqrt(cx * cx + cy * cy + cz * cz);
     }
 
     /// Generate pseudo points from a single MeshChunk
     inline void generatePseudoPointsFromMesh(
         const chunkbim::MeshChunk& mesh,
-        float gridSize,
-        std::vector<std::array<float, 3>>& outPoints,
+        double gridSize,
+        std::vector<std::array<double, 3>>& outPoints,
         std::vector<size_t>& faceIdx,
-        std::vector<std::array<float, 3>>& facePts,
-        std::vector<std::array<float, 3>>& outNormals)
+        std::vector<std::array<double, 3>>& facePts,
+        std::vector<std::array<double, 3>>& outNormals)
     {
         if (mesh.faces.empty())
             return;
 
-        float gridSizeSquared = gridSize * gridSize;
+        double gridSizeSquared = gridSize * gridSize;
 
         for (size_t f = 0; f < mesh.faces.size(); ++f) {
             const auto& face = mesh.faces[f];
-            
-            std::array<float, 3> fPt;
+
+            std::array<double, 3> fPt;
             fPt[0] = face.vertices[0][0];
             fPt[1] = face.vertices[0][1];
             fPt[2] = face.vertices[0][2];
 
             facePts.push_back(fPt);
-            
+
             /// Normal from face
-            std::array<float, 3> fNormal;
+            std::array<double, 3> fNormal;
             fNormal[0] = face.normal[0];
             fNormal[1] = face.normal[1];
             fNormal[2] = face.normal[2];
@@ -81,8 +81,8 @@ namespace icp {
             const auto& v2 = face.vertices[2];
 
             /// Calculate triangle area
-            float area = triangleArea(v0, v1, v2);
-            if (area < 1e-10f) continue;  /// Skip degenerate triangles
+            double area = triangleArea(v0, v1, v2);
+            if (area < 1e-10) continue;  /// Skip degenerate triangles
 
             /// Number of points to generate based on area and grid size
             int numPoints = static_cast<int>(std::ceil(area / gridSizeSquared));
@@ -90,19 +90,19 @@ namespace icp {
 
             /// Generate points using barycentric coordinates
             /// Use deterministic pattern for reproducibility
-            int gridDim = static_cast<int>(std::ceil(std::sqrt(static_cast<float>(numPoints))));
-            
+            int gridDim = static_cast<int>(std::ceil(std::sqrt(static_cast<double>(numPoints))));
+
             for (int i = 0; i < gridDim; ++i) {
                 for (int j = 0; j < gridDim - i; ++j) {
                     /// Barycentric coordinates
-                    float u = (i + 0.5f) / gridDim;
-                    float v = (j + 0.5f) / gridDim;
-                    float w = 1.0f - u - v;
+                    double u = (i + 0.5) / gridDim;
+                    double v = (j + 0.5) / gridDim;
+                    double w = 1.0 - u - v;
 
-                    if (w < 0.0f) continue;  /// Outside triangle
+                    if (w < 0.0) continue;  /// Outside triangle
 
                     /// Interpolate position
-                    std::array<float, 3> point;
+                    std::array<double, 3> point;
                     point[0] = u * v0[0] + v * v1[0] + w * v2[0];
                     point[1] = u * v0[1] + v * v1[1] + w * v2[1];
                     point[2] = u * v0[2] + v * v1[2] + w * v2[2];
@@ -117,11 +117,11 @@ namespace icp {
     /// Generate pseudo points from multiple MeshChunks
     inline void generatePseudoPoints(
         const std::vector<chunkbim::MeshChunk>& chunks,
-        float gridSize,
-        std::vector<std::array<float, 3>>& outPoints,
+        double gridSize,
+        std::vector<std::array<double, 3>>& outPoints,
         std::vector<size_t>& faceIdx,
-        std::vector<std::array<float, 3>>& facePts,
-        std::vector<std::array<float, 3>>& outNormals)
+        std::vector<std::array<double, 3>>& facePts,
+        std::vector<std::array<double, 3>>& outNormals)
     {
         outPoints.clear();
         faceIdx.clear();
@@ -149,10 +149,10 @@ namespace icp {
     }
 
     /// Convert PointCloudChunk to ICP format
-    inline std::vector<std::array<float, 3>> convertPointCloudToIcp(
+    inline std::vector<std::array<double, 3>> convertPointCloudToIcp(
         const std::vector<chunkpc::PointCloudChunk>& chunks)
     {
-        std::vector<std::array<float, 3>> result;
+        std::vector<std::array<double, 3>> result;
 
         /// Count total points
         size_t totalPoints = 0;
